+++ bin/flows/build-chain-single.js
@@ -42,9 +42,10 @@ async function execute(
  * @param {Object} env proces.env
  * @param {Object} the JSON object for the event data
  */
-async function executeFromEvent(token, octokit, env, eventData) {
+async function executeFromEvent(token, octokit, env, eventData, options = {}) {
   await execute(token, octokit, env, eventData, undefined, {
-    isArchiveArtifacts: true
+    isArchiveArtifacts: true,
+    ...options
   });
 }