+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Prints system git version
 - [issue 203] Prints execution summary
+- additional-flags input
 
 ## Bugs